requirements.txt
Django==2.0.6
django-bootstrap3==10.0.1
django-multiselectfield==0.1.8
pytz==2018.4
